change epochs type

+++ train.py
@@ -33,7 +33,7 @@ parser.add_argument('--l_adv', default=1, type=float)
 parser.add_argument('--l_rec', default=100, type=float)
 parser.add_argument('--l_pix', default=10, type=float)
 parser.add_argument('--l_percept', default=10, type=float)
-parser.add_argument('--epochs', default=20, type=float)
+parser.add_argument('--epochs', default=20, type=int)
 args = parser.parse_args()
 
 np.random.seed(12)
